refactor: V1 API 제거 (#332)
* refactor: deprecate about v1 api

* refactor: remove deprecated academics v1 api

* refactor: remove unused dto

* refactor: move admin apis to v2

* refactor: remove v1 admission api, refactor service method name

* refactor: move conference api to v2

* refactor: move internal api to v2

* refactor: move main api to v2

* refactor: remove deprecated v1 api for member

* refactor: move news, notice, seminar apis to v2

* refactor: move recruit api to v2

* refactor: remove deprecated research v1 api

* refactor: move reservation api to v2

* refactor: move user api to v2

* lint: fix ktlint

* fix: restore falsly deleted GroupedCourseDto
