--- src/main/kotlin/com/wafflestudio/csereal/core/about/dto/AboutRequest.kt
@@ -1,7 +1,0 @@
-package com.wafflestudio.csereal.core.about.dto
-
-data class AboutRequest(
-    val postType: String,
-    val language: String,
-    val description: String
-)
