--- src/main/kotlin/com/wafflestudio/csereal/core/about/dto/FutureCareersRequest.kt
@@ -1,8 +1,0 @@
-package com.wafflestudio.csereal.core.about.dto
-
-data class FutureCareersRequest(
-    val language: String,
-    val description: String,
-    val stat: List<FutureCareersStatDto>,
-    val companies: List<FutureCareersCompanyDto>
-)
