+++ src/main/kotlin/com/wafflestudio/csereal/core/admissions/api/v2/AdmissionsController.kt
@@ -1,17 +1,15 @@
 package com.wafflestudio.csereal.core.admissions.api.v2
 
 import com.wafflestudio.csereal.common.aop.AuthenticatedStaff
+import com.wafflestudio.csereal.common.enums.LanguageType
 import com.wafflestudio.csereal.core.admissions.api.req.UpdateAdmissionReq
 import com.wafflestudio.csereal.core.admissions.api.res.GroupedAdmission
 import com.wafflestudio.csereal.core.admissions.service.AdmissionsService
 import com.wafflestudio.csereal.core.admissions.type.AdmissionsMainType
 import com.wafflestudio.csereal.core.admissions.type.AdmissionsPostType
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PathVariable
-import org.springframework.web.bind.annotation.PutMapping
-import org.springframework.web.bind.annotation.RequestBody
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RestController
+import jakarta.validation.Valid
+import jakarta.validation.constraints.Positive
+import org.springframework.web.bind.annotation.*
 
 @RequestMapping("/api/v2/admissions")
 @RestController
@@ -28,6 +26,8 @@ class AdmissionsController(
         return admissionsService.readGroupedAdmission(mainType, postType)
     }
 
+    // TODO: Add Create, Delete Admission Pair API if needed
+
     @AuthenticatedStaff
     @PutMapping("/{mainTypeStr}/{postTypeStr}")
     fun updateAdmission(
@@ -37,6 +37,34 @@ class AdmissionsController(
     ) {
         val mainType = AdmissionsMainType.fromJsonValue(mainTypeStr)
         val postType = AdmissionsPostType.fromJsonValue(postTypeStr)
-        admissionsService.updateAdmission(mainType, postType, updateAdmissionReq)
+        admissionsService.updateGroupedAdmission(mainType, postType, updateAdmissionReq)
     }
+
+    @GetMapping("/search/top")
+    fun searchTopAdmissions(
+        @RequestParam(required = true) keyword: String,
+        @RequestParam(required = true, defaultValue = "ko") language: String,
+        @RequestParam(required = true) @Valid @Positive number: Int,
+        @RequestParam(required = false, defaultValue = "30") @Valid @Positive amount: Int
+    ) = admissionsService.searchTopAdmission(
+        keyword,
+        LanguageType.makeStringToLanguageType(language),
+        number,
+        amount
+    )
+
+    @GetMapping("/search")
+    fun searchPageAdmissions(
+        @RequestParam(required = true) keyword: String,
+        @RequestParam(required = true, defaultValue = "ko") language: String,
+        @RequestParam(required = true) @Valid @Positive pageSize: Int,
+        @RequestParam(required = true) @Valid @Positive pageNum: Int,
+        @RequestParam(required = false, defaultValue = "30") @Valid @Positive amount: Int
+    ) = admissionsService.searchPageAdmission(
+        keyword,
+        LanguageType.makeStringToLanguageType(language),
+        pageSize,
+        pageNum,
+        amount
+    )
 }
